requirements.txt
# Core Streamlit and Web Framework
streamlit==1.40.1
streamlit-option-menu==0.4.0

# PDF Processing
PyPDF2==3.0.1

# AI and Machine Learning
sentence-transformers==3.3.1
google-generativeai==0.8.3
chromadb==0.5.23

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Web Search (Optional - for agent capabilities)
requests==2.32.3
beautifulsoup4==4.12.3

# Environment Management
python-decouple==3.8

# Additional utilities
plotly==5.24.1